Added wikisync info.
