+++ .gitignore
@@ -1,0 +1,5 @@
+.env
+wikisync.log
+wikisync.cookies
+upload_map.yml
+temp
