--- CNAME
@@ -1,1 +1,0 @@
-igem.kylelierer.com
