+++ requirements.txt
@@ -1,0 +1,2 @@
+igem_wikisync==1.0.0
+python-dotenv==0.17.1
