+++ site/assets/css/igem-remover.css
@@ -38,3 +38,11 @@ html {
 #globalWrapper {
     font-size: 100% !important;
 }
+
+#top_menu_14 {
+    height: 18px;
+}
+
+#top_menu_under {
+    height: 4px;
+}
